A2 (tail): lift TileCache + render keys from app into ui
Move the TileCache struct and its two staleness-key enums (ModalFocusKey,
ScrollContentKey) out of app into ui::tiles — the cache is now ui-owned with
no App reference, so a ui-only harness (D) could render without app. The keys
reference core types (Settings, LogLevelOverride, Screen) only, so ui stays a
dependency leaf. app/tiles.rs deleted; render methods stay on impl App taking
a &TileCache param (moving them onto impl TileCache needs the whole app::view
rasterization layer relocated — a separate project, see the doc).
